CreatingTasks/requirements.txt
# Django
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Real-time features
channels==4.0.0
channels-redis==4.1.0
asgiref==3.7.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Database
psycopg2-binary==2.9.7

# Telegram bot
aiogram==3.2.0
aiohttp==3.9.1

# Authentication
django-allauth==0.54.0
django-rest-framework-simplejwt==5.3.0

# Development (Windows compatible)
python-dotenv==1.0.0
pillow==10.0.1

# Async
asyncio==3.4.3